build: Move the pnglibconf files and scripts to scripts/pnglibconf/
Considering that pnglibconf is a whole mini-system for configuring and
building libpng, it should be placed in its own dedicated subdirectory.
Any future additions, including tests and documentation, should go into
this subdirectory.

One exception: pnglibconf.h.prebuilt is moved to the top-level source
directory, where pnglibconf.h would be. This follows the organization
pattern typically used with the "*.in" files, for example.

All references to the pnglibconf artifacts, found in machine-readable
code and in human-readable comments, are updated accordingly.
