+++ scripts/cmake/gensrc.cmake.in
@@ -1,7 +1,7 @@
 # gensrc.cmake.in
 # Generate source files with awk, based upon the automake logic.
 
-# Copyright (c) 2022-2024 Cosmin Truta
+# Copyright (c) 2022-2025 Cosmin Truta
 # Copyright (c) 2016 Glenn Randers-Pehrson
 # Written by Roger Leigh, 2016
 #
@@ -21,24 +21,27 @@ set(DFA_XTRA "@DFA_XTRA@")
 set(PNG_PREFIX "@PNG_PREFIX@")
 set(PNGLIB_VERSION "@PNGLIB_VERSION@")
 
-if(OUTPUT STREQUAL "scripts/pnglibconf.c")
-  # Generate scripts/pnglibconf.c
+if(OUTPUT STREQUAL "scripts/pnglibconf/pnglibconf.c")
+  # Generate scripts/pnglibconf/pnglibconf.c
 
   file(REMOVE "${BINDIR}/pnglibconf.tf6" "${BINDIR}/pnglibconf.tf7")
 
   execute_process(COMMAND "${CMAKE_COMMAND}" -E echo "com ${PNGLIB_VERSION} STANDARD API DEFINITION"
-                  COMMAND "${AWK}" -f "${SRCDIR}/scripts/options.awk"
-                          "out=pnglibconf.tf6" "logunsupported=1" "version=search"
+                  COMMAND "${AWK}" -f "${SRCDIR}/scripts/pnglibconf/options.awk"
+                          "out=pnglibconf.tf6"
+                          "logunsupported=1"
+                          "version=search"
                           "${SRCDIR}/pngconf.h" "-"
-                          "${SRCDIR}/scripts/pnglibconf.dfa"
+                          "${SRCDIR}/scripts/pnglibconf/pnglibconf.dfa"
                   WORKING_DIRECTORY "${BINDIR}"
                   RESULT_VARIABLE AWK_FAIL)
   if(AWK_FAIL)
     message(FATAL_ERROR "Failed to generate pnglibconf.tf6")
   endif()
 
-  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/options.awk"
-                  "out=pnglibconf.tf7" "pnglibconf.tf6"
+  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/pnglibconf/options.awk"
+                  "out=pnglibconf.tf7"
+                  "pnglibconf.tf6"
                   WORKING_DIRECTORY "${BINDIR}"
                   RESULT_VARIABLE AWK_FAIL)
   if(AWK_FAIL)
@@ -47,25 +50,29 @@ if(OUTPUT STREQUAL "scripts/pnglibconf.c")
 
   file(REMOVE "pnglibconf.tf6")
   file(MAKE_DIRECTORY "${BINDIR}/scripts")
-  file(RENAME "pnglibconf.tf7" "${BINDIR}/scripts/pnglibconf.c")
+  file(RENAME "pnglibconf.tf7" "${BINDIR}/scripts/pnglibconf/pnglibconf.c")
 
 elseif(OUTPUT STREQUAL "pnglibconf.c")
   # Generate pnglibconf.c
 
   file(REMOVE "${BINDIR}/pnglibconf.tf4" "${BINDIR}/pnglibconf.tf5")
 
-  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/options.awk"
-                  out=pnglibconf.tf4 version=search
-                  ${SRCDIR}/pngconf.h ${SRCDIR}/scripts/pnglibconf.dfa
-                  ${SRCDIR}/pngusr.dfa ${DFA_XTRA}
+  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/pnglibconf/options.awk"
+                  "out=pnglibconf.tf4"
+                  "version=search"
+                  "${SRCDIR}/pngconf.h"
+                  "${SRCDIR}/scripts/pnglibconf/pnglibconf.dfa"
+                  "${SRCDIR}/pngusr.dfa"
+                  ${DFA_XTRA}
                   WORKING_DIRECTORY "${BINDIR}"
                   RESULT_VARIABLE AWK_FAIL)
   if(AWK_FAIL)
     message(FATAL_ERROR "Failed to generate pnglibconf.tf4")
   endif()
 
-  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/options.awk"
-                  out=pnglibconf.tf5 pnglibconf.tf4
+  execute_process(COMMAND "${AWK}" -f "${SRCDIR}/scripts/pnglibconf/options.awk"
+                  "out=pnglibconf.tf5"
+                  "pnglibconf.tf4"
                   WORKING_DIRECTORY "${BINDIR}"
                   RESULT_VARIABLE AWK_FAIL)
   if(AWK_FAIL)
@@ -87,9 +94,11 @@ elseif(OUTPUT STREQUAL "pnglibconf.h")
                              s==0{prev=\$0}
                              s==1{print \"#define\", \$1, \"${PNG_PREFIX}\" \$1}
                              s==2{print \"#define ${PNG_PREFIX}png_\" \$1, \"PNG_\" \$1}
-                             END{print prev}" s=0 pnglibconf.out s=1 "${BINDIR}/scripts/prefix.out"
+                             END{print prev}"
+                             s=0 "pnglibconf.out"
+                             s=1 "${BINDIR}/scripts/pnglibconf/prefix.out"
                              s=2 "${SRCDIR}/scripts/macro.lst"
-                    OUTPUT_FILE pnglibconf.tf8
+                    OUTPUT_FILE "pnglibconf.tf8"
                     RESULT_VARIABLE AWK_FAIL)
     if(AWK_FAIL)
       message(FATAL_ERROR "Failed to generate pnglibconf.tf8")
@@ -115,7 +124,7 @@ elseif(OUTPUT STREQUAL "pngprefix.h")
 
     execute_process(COMMAND "${AWK}"
                             "{print \"#define\", \$1, \"${PNG_PREFIX}\" \$1}"
-                            "${BINDIR}/scripts/intprefix.out"
+                            "${BINDIR}/scripts/pnglibconf/intprefix.out"
                     OUTPUT_FILE "pngprefix.tf1"
                     RESULT_VARIABLE AWK_FAIL)
     if(AWK_FAIL)
@@ -127,13 +136,13 @@ elseif(OUTPUT STREQUAL "pngprefix.h")
     file(WRITE "${BINDIR}/${OUTPUT}" "/* No libpng symbol prefix configured. */")
   endif()
 
-elseif(OUTPUT STREQUAL "scripts/pnglibconf.h.prebuilt")
-  # Generate scripts/pnglibconf.h.prebuilt (fails build)
+elseif(OUTPUT STREQUAL "pnglibconf.h.prebuilt")
+  # Generate pnglibconf.h.prebuilt (fails build)
 
-  message(STATUS "Attempting to build scripts/pnglibconf.h.prebuilt")
+  message(STATUS "Attempting to build pnglibconf.h.prebuilt")
   message(STATUS "This is a machine generated file, but if you want to make")
   message(STATUS "a new one simply build the 'png_genfiles' target, and copy")
-  message(STATUS "scripts/pnglibconf.out to scripts/pnglibconf.h.prebuilt")
+  message(STATUS "scripts/pnglibconf/pnglibconf.out to pnglibconf.h.prebuilt")
   message(STATUS "AND set PNG_ZLIB_VERNUM to 0 (you MUST do this)")
   message(FATAL_ERROR "Stopping build")
