+++ scripts/makefile.emcc
@@ -36,8 +36,8 @@ PNGTEST_LDFLAGS = --preload-file=pngtest.png
 LIBS = -lz -lm
 
 # Pre-built configuration
-# See scripts/pnglibconf.mak for more options
-PNGLIBCONF_H_PREBUILT = scripts/pnglibconf.h.prebuilt
+# See scripts/pnglibconf/pnglibconf.mak for more options
+PNGLIBCONF_H_PREBUILT = pnglibconf.h.prebuilt
 
 # File lists
 OBJS = png.o pngerror.o pngget.o pngmem.o pngpread.o pngread.o \
