+++ scripts/makefile.ibmc
@@ -29,8 +29,8 @@ CFLAGS = -Mc -O2 -W3
 LDFLAGS =
 
 # Pre-built configuration
-# See scripts/pnglibconf.mak for more options
-PNGLIBCONF_H_PREBUILT = scripts/pnglibconf.h.prebuilt
+# See scripts/pnglibconf/pnglibconf.mak for more options
+PNGLIBCONF_H_PREBUILT = pnglibconf.h.prebuilt
 
 # File extensions
 O=.obj
