+++ scripts/pnglibconf/intprefix.c
@@ -17,4 +17,4 @@
         PNG_DFN "@" name "@"
 
 #define PNGPREFIX_H /* self generation */
-#include "../pngpriv.h"
+#include "../../pngpriv.h"
