+++ scripts/pnglibconf/vers.c
@@ -1,4 +1,4 @@
-/* sym.c - define format of libpng.sym
+/* vers.c - define format of libpng.vers
  *
  * Copyright (c) 2011-2014 Glenn Randers-Pehrson
  *
@@ -8,6 +8,10 @@
  */
 
 #define PNG_EXPORTA(type, name, args, attributes)\
-        PNG_DFN "@" SYMBOL_PREFIX "@@" name "@"
+        PNG_DFN " @" SYMBOL_PREFIX "@@" name "@;"
 
-#include "../png.h"
+PNG_DFN "@" PNGLIB_LIBNAME "@ {global:"
+
+#include "../../png.h"
+
+PNG_DFN "local: *; };"
